# perf: Add a simple benchmark utility for prototype methods