This uses a proxy to intercept and time function invocations within a given object.

Example usage in GWL:

```js
imports.perf.core.benchmarkPrototype(AppGroup, 4);
```

Output:

```
Cjs-Message: 09:28:03.887: JS LOG: [LookingGlass/info] AppGroup.onFocusChange: 6ms (MAX: 26ms AVG: 5.63ms)
Cjs-Message: 09:28:03.887: JS LOG: [LookingGlass/info] AppGroup.onFocusWindowChange: 6ms (MAX: 26ms AVG: 5.63ms)
Cjs-Message: 09:28:03.890: JS LOG: [LookingGlass/info] AppGroup.onFocusChange: 4ms (MAX: 26ms AVG: 5.62ms)
Cjs-Message: 09:28:03.891: JS LOG: [LookingGlass/info] AppGroup.onFocusWindowChange: 4ms (MAX: 26ms AVG: 5.61ms)
Cjs-Message: 09:28:08.119: JS LOG: [LookingGlass/info] AppGroup.windowAdded: 30ms (MAX: 30ms AVG: 5.77ms)
Cjs-Message: 09:28:11.324: JS LOG: [LookingGlass/info] AppGroup.onFocusChange: 4ms (MAX: 30ms AVG: 5.75ms)
```